feat: Store a dict of moves from every board pos
Piece actions depend on the available moves from their board position.
Redundant move evaluations are avoided by storing precalculated values.
Moves are stored as a dictionary with a key for every board position.

From a position, moves can be made in up to four direction.
These moves can be either to an adjacent empty position or a jump.
Each move direction has a tuple of the adjacent move and jump move.
Regular pieces can either move upward or downward but not both.
Therefore, separate lists of up and down moves are stored as a tuple.