fix: Make piece on last row of test_board a king
Negative valued pieces initially move top to bottom across the board.
When pieces arrive at the opposite end of the board they become kings.
Kings are denoted with an absolute value greater than 1.

In the test_board, the -1 valued piece in the last row should be a king.
Therefore, its value is changed to -2.