feat: Add method to evaluate the game outcome
When a game has ended, it is necessary to determine the winner.
Added a method to return a player reward tuple for a checkers state.
If the game has ended and a player wins, their reward value is +1.
If the game has ended and a player loses, their reward value is -1.
Otherwise, the reward for both players is 0.

resolves #9