feat: Add continuation of jump sequence
In one turn a player can use a piece to make a sequence of jumps.
After a jump, the piece used can make additional jumps.
Each jump results in a new state.
The state at the befinning of a players turn has all legal action.
A state following a jump only yields jumps for the previous used piece.

resolves #13