style: Differentiate "piece" from "postion"
Each checker piece occupies a unique position on the board.
Therefore, position tuples are often used to represent pieces.
Every piece has a position, but not every board position has a piece.
The code used the names "piece" and "position" interchangably.
This change is to make the naming scheme more consistent.
 - "piece" is used for positions that always represent pieces
 - "position" is used for spaces that do not necessarily contain pieces