bug: Player never starts turn in a jump state
After a player makes a jump, the resulting state is a jump state.
If no further jumps are available, the other player becomes active.
When the new player starts their turn, it should not be a jump state.
Previously that was the case, this commit fixes that bug.
If that player does not have actions, the first player becomes active.
In that case the player is no longer restricted to continued jump moves.
And if there are still no available actions, the game is a draw.

resolves #17