refactor: Add getter methods for public variables
Some instatnce variables should be publically visible.
`actions` and `is_terminal` were both treated as public variables.
These variables are not meant to be publically modified though.
To achieve this, the variables are marked for internal use only.
This is achieved by adding the `_` prefix.
The values are then made externally accessible by adding getter methods.
The active player is also included as one of these variables.

resolves #