feat: Create base State class
CheckersState is a specific instance of a game state.
Agents and game controllers can be generic for different types of games.
To test them, it can be useful to make simpler/tailored states.
To make other games, it is useful to derive from a shared base class.
The base class defines the expected interfaces of a game state.