feat: Create GameController
A game controller instances coordinates the execution of a game.
It is initiailzed with a starting game state and a list of agents.
The play_game() method gets agent actions and evolves game states.
Once a terminal state is reached, the game outcome is returned.
A full game of checkers is tested using a GameController.

resolves #20