feat: Add labels to checker board visualization
For manual play, a person needs to understand the actions presented.
The tuple of coordinates is not the most human readable format.
Spaces on a chess board are often labeled with letter number pair.
This is a more familiar and understandable labeling scheme.
These labels are added to the board visualization.