feat: Add more functionality to DummyState
DummyState is a subclass of State designed for testing.
The only functional method it implemented as get_actions().
For more complex tests, the other state methods need to be implemented.
Available actions and their results are represented state subtree dict.
The actions are the keys, the results are the corresponding values.
A state is terminal when it does not has an empty subtree.
The active player and score can also be defined in a dummy state.