feat: Add BFSAgent
This agent performs a breadth first search on the tree of future states.
Like the Greedy Agent it is initialized with a function to score states.
It also takes an argument for the maximum depth of the search.
The default is to search 3 states in the future.
It then assumes each active acent acts to maximize their value.

resolves #25